+++ src/feditest/nodedrivers/mastodon/ubos.py
@@ -21,21 +21,7 @@ from feditest.nodedrivers.mastodon import (
     USERID_ACCOUNT_FIELD,
     USERID_NON_EXISTING_ACCOUNT_FIELD
 )
-from feditest.protocols import (
-    Account,
-    NonExistingAccount,
-    AccountManager,
-    DefaultAccountManager,
-    Node,
-    NodeConfiguration,
-    APP_PAR,
-    APP_VERSION_PAR,
-    HOSTNAME_PAR,
-)
-from feditest.registry import registry_singleton
-from feditest.reporting import error, trace
-from feditest.testplan import TestPlanConstellationNode, TestPlanNodeAccountField, TestPlanNodeNonExistingAccountField, TestPlanNodeParameterMalformedError
-from feditest.ubos import (
+from feditest.nodedrivers.ubos import (
     UbosNodeConfiguration,
     UbosNodeDeployConfiguration,
     UbosNodeDriver,
@@ -52,6 +38,20 @@ from feditest.ubos import (
     TLSCERT_PAR,
     TLSKEY_PAR
 )
+from feditest.protocols import (
+    Account,
+    NonExistingAccount,
+    AccountManager,
+    DefaultAccountManager,
+    Node,
+    NodeConfiguration,
+    APP_PAR,
+    APP_VERSION_PAR,
+    HOSTNAME_PAR,
+)
+from feditest.registry import registry_singleton
+from feditest.reporting import error, trace
+from feditest.testplan import TestPlanConstellationNode, TestPlanNodeAccountField, TestPlanNodeNonExistingAccountField, TestPlanNodeParameterMalformedError
 
 
 class MastodonUbosNodeConfiguration(UbosNodeDeployConfiguration, NodeWithMastodonApiConfiguration):
